factor out business logic (i.e. non router) into components/Fencing
